user total assets procedure
